requirements.txt
flask
torch
open_clip_torch
scikit-learn
numpy
pandas
pillow